LayoutTests/fast/forms/month/input-valueasdate-expected.txt

Test HTMLInputElement::valueAsDate binding.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Unsuppported type:
PASS input.valueAsDate is null
PASS input.valueAsDate = date threw exception Error: InvalidStateError: DOM Exception 11.

Supported type:
PASS valueAsDate != null is true
PASS typeof valueAsDate is "object"
PASS valueAsDate.constructor.name is "Date"
Sets an Epoch Date:
PASS input.value is "1970-01"
PASS input.valueAsDate.getTime() is 0
Sets a number 0:
PASS input.value is "1970-01"
PASS input.valueAsDate.getTime() is 0
Sets other types:
PASS input.value is ""
PASS input.value is ""
PASS input.value is ""
PASS successfullyParsed is true

TEST COMPLETE
